add readings for my previous posts

+++ tools/gen_tts.js
@@ -6,7 +6,7 @@ const fm = require('front-matter');
 const path = require('path')
 const audioconcat = require('audioconcat');
 
-const { TTS } = require('./tts');
+const { TTS } = require('@tyrchen/xftts');
 const env = process.env;
 const app = env.APP || '1';
 const appId = env[`XF_TTS_ID${app}`];
@@ -21,14 +21,13 @@ async function gen_mp3(text, filename) {
   text = text.replace(/”|「|」|"|_|《|》|>/g, '');
   // console.log(text);
   try {
-    const tts = new TTS(appId, apiKey, apiSecret, {vcn: 'x2_xiaoyuan', speed: 70});
+    const tts = new TTS(appId, apiKey, apiSecret, { vcn: 'x2_xiaoyuan', speed: 70 });
     await tts.generate(text, filename);
     console.log('generated: ', filename);
   } catch (error) {
     console.error(error);
   }
 }
-
 
 async function extract(src) {
   const content = await fs.readFile(src, 'utf8');
@@ -51,35 +50,41 @@ async function main() {
   const outPath = path.join(path.dirname(argv.i), 'assets');
   const outName = path.basename(argv.i).split('-')[0];
   const data = await extract(argv.i);
-
-  let files = [];
-  for (const [i, item] of data.entries()) {
-    let outFile = path.join(outPath, `${outName}-${i}.mp3`);
-    await gen_mp3(item, outFile);
-    files.push(outFile);
-  }
-
   const output = path.join(outPath, `${outName}.mp3`);
-  setTimeout(() => 
-  audioconcat(files)
-    .concat(output)
-    .on('start', function (command) {
-      console.log('ffmpeg process started:', command);
 
-    })
-    .on('error', function (err, _stdout, stderr) {
-      console.error('Error:', err);
-      console.error('ffmpeg stderr:', stderr);
-      process.exit(-1);
-    })
-    .on('end', async function () {
-      for (var f of files) {
-        await fs.unlink(f);
-      }
-      console.error('Audio created in:', output);
-      process.exit(0);
-    })
-  , 1000);
+  if (data.length == 1) {
+    await gen_mp3(data[0], output);
+    console.error('Audio created in:', output);
+    process.exit(0);
+  } else {
+    let files = [];
+    for (const [i, item] of data.entries()) {
+      let outFile = path.join(outPath, `${outName}-${i}.mp3`);
+      await gen_mp3(item, outFile);
+      files.push(outFile);
+    }
+
+    setTimeout(() =>
+      audioconcat(files)
+        .concat(output)
+        .on('start', function (command) {
+          console.log('ffmpeg process started:', command);
+
+        })
+        .on('error', function (err, _stdout, stderr) {
+          console.error('Error:', err);
+          console.error('ffmpeg stderr:', stderr);
+          process.exit(-1);
+        })
+        .on('end', async function () {
+          for (var f of files) {
+            await fs.unlink(f);
+          }
+          console.error('Audio created in:', output);
+          process.exit(0);
+        })
+      , 1000);
+  }
 }
 
 main()
